apps/offer-scraper/requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-core==0.1.10
langgraph==0.0.20
requests==2.31.0
python-dotenv==1.0.0
